refactoring code successfully
