--- src/main/java/entities/Cor.java
@@ -1,11 +1,0 @@
-package entities;
-
-public enum Cor {
-	
-	BLUE,
-	BLACK,
-	GRAY,
-	PURPLE,
-	RED,
-	YELLOW
-}
