+++ src/test/java/br/nttdata/pageFactories/HomePageFactory.java
@@ -1,14 +1,11 @@
 package br.nttdata.pageFactories;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-
-import exceptions.HomePageException;
 
 public class HomePageFactory extends BasePageFactory{
 
@@ -28,6 +25,7 @@ public class HomePageFactory extends BasePageFactory{
 		btnSpecialOffer.click();
 	}
 	
+	
 	public void clicarEmSeeOffer() {
 		
 		((JavascriptExecutor) driver).executeScript("scroll(0,300)");
@@ -35,11 +33,4 @@ public class HomePageFactory extends BasePageFactory{
 		actions.moveToElement(btnSeeOffer).click().perform();
 		wait.until(ExpectedConditions.elementToBeClickable(btnSeeOffer)).click();
 	}
-
-	
-
-	
-	
-	
-
 }
