--- src/test/java/br/nttdata/steps/ProdSteps.java
@@ -1,82 +1,0 @@
-//package br.nttdata.steps;
-//
-//import io.cucumber.java.pt.Dado;
-//import io.cucumber.java.pt.E;
-//import io.cucumber.java.pt.Então;
-//import io.cucumber.java.pt.Quando;
-//
-//public class ProdSteps {
-//	
-//	@Dado("que acessei o site advantageonlineshopping")
-//	public void queAcesseiOSiteAdvantageonlineshopping() {
-//	}
-//	    
-//	@E("no menu clicar na opção Special Offer")
-//	public void noMenuClicarNaOpçãoSpecialOffer() {
-//	    
-//	}
-//	@Quando("clicar no botão See Offer")
-//	public void clicarNoBotãoSeeOffer() {
-//	
-//	}
-//	   
-//	@Então("vou validar as especificações do produto")
-//	public void vouValidarAsEspecificaçõesDoProduto() {
-//	   
-//	}
-//	
-//	@E("alterar a cor do produto")
-//	public void alterarACorDoProduto() {
-//	    
-//	}
-//	@Quando("clicar no botão Add to cart")
-//	public void clicarNoBotãoAddToCart() {
-//	   
-//	}
-//	@Então("o produto foi adicionado ao carrinho")
-//	public void oProdutoFoiAdicionadoAoCarrinho() {
-//	   
-//	}
-//
-//	
-//	@E("pesquisar o produto clicando no ícone de lupa")
-//	public void pesquisarOProdutoClicandoNoÍconeDeLupa() {
-//		
-//	}
-//	
-//	@E("selecionar o produto pesquisado")
-//	public void selecionarOProdutoPesquisado() {
-//	  
-//	}
-//
-//	@E("alterar a quantidade de produtos")
-//	public void alterarAQuantidadeDeProdutos() {
-//	    
-//	}
-//	
-//	@E("acessar a página de checkout")
-//	public void acessarAPáginaDeCheckout() {
-//	   
-//	}
-//	@Então("vou validar que a soma dos preços corresponde ao total da página de checkout")
-//	public void vouValidarQueASomaDosPreçosCorrespondeAoTotalDaPáginaDeCheckout() {
-//	    
-//	}
-//	
-//	@E("remover o produto do carrinho de compras")
-//	public void removerOProdutoDoCarrinhoDeCompras() {
-//	}
-//	
-//	@Então("o carinho de compras estará vazio")
-//	public void oCarinhoDeComprasEstaráVazio() {
-//	    
-//	}
-//
-//
-//
-//
-//
-//
-//
-//
-//}
