--- src/test/resources/features/produto1.feature
@@ -1,42 +1,0 @@
-# language: pt
-#
-#Funcionalidade: Validações do produto
-#
-#Como um usuário
-#Gostaria de incluir e excluir produtos do carrinho
-#Para que eu posso validar as informações do carrinho
-#	
-#
-#Cenario: Validar especificações do produto
-#	Dado que acessei o site advantageonlineshopping
-#	E no menu clicar na opção Special Offer
-#	Quando clicar no botão See Offer
-#	Então vou validar as especificações do produto
-#	
-#Cenario: Validar alteração de cor do produto no carrinho
-#	Dado que acessei o site advantageonlineshopping
-#	E no menu clicar na opção Special Offer
-#	Quando clicar no botão See Offer
-#	E alterar a cor do produto
-#	Quando clicar no botão Add to cart
-#	Então o produto foi adicionado ao carrinho
-#	
-#Cenario: Validar página de checkout
-#	Dado que acessei o site advantageonlineshopping
- #	E pesquisar o produto clicando no ícone de lupa
- #	E selecionar o produto pesquisado
- #	E alterar a cor do produto
- #	E alterar a quantidade de produtos
- #	Quando clicar no botão Add to cart
- #	E acessar a página de checkout
- #	Então vou validar que a soma dos preços corresponde ao total da página de checkout
- #	
- #Cenario: Remover produto do carrinho de compras
-#	Dado que acessei o site advantageonlineshopping
- #	E no menu clicar na opção Special Offer
- #	Quando clicar no botão See Offer
- #	E remover o produto do carrinho de compras
- #	Então o carinho de compras estará vazio
- #	
- #	
-#
